StorageConverterStorageBinaryToStorageValueShared was StorageConverterStorageBinaryToStorageValue
